performance_tester/requirements.txt
flask==3.0.0
fpdf2==2.7.6
matplotlib==3.8.2
python-dotenv==1.0.0
requests==2.31.0